@@ -1,7 +1,7 @@
 {
   "name": "kepler.gl",
   "author": "Shan He <shan@uber.com>",
-  "version": "1.1.9",
+  "version": "1.1.11",
   "description": "kepler.gl is a webgl based application to visualize large scale location data in the browser",
   "license": "MIT",
   "main": "dist/index.js",
@@ -94,6 +94,7 @@
     "lodash.curry": "^4.1.1",
     "lodash.debounce": "^4.0.8",
     "lodash.flattendeep": "^4.2.0",
+    "lodash.isequal": "^4.5.0",
     "lodash.memoize": "^4.1.2",
     "lodash.pick": "^4.4.0",
     "lodash.throttle": "^4.1.1",
@@ -104,14 +105,14 @@
     "pbf": "^3.1.0",
     "prop-types": "^15.6.0",
     "react-color": "^2.17.3",
-    "react-file-drop": "^0.1.9",
     "react-json-pretty": "^1.7.9",
+    "react-lifecycles-compat": "^3.0.4",
     "react-map-gl": "^5.0.3",
     "react-markdown": "^4.0.6",
     "react-modal": "^3.8.1",
     "react-onclickoutside": "^6.7.1",
     "react-palm": "^3.1.2",
-    "react-redux": "^5.1.1",
+    "react-redux": "^7.1.3",
     "react-sortable-hoc": "^1.8.3",
     "react-tooltip": "^3.10.0",
     "react-virtualized": "^9.21.1",
@@ -203,8 +204,8 @@
     "webpack-stats-plugin": "^0.2.1"
   },
   "peerDependencies": {
-    "react": "16.x",
-    "react-dom": "16.x",
+    "react": "^16.3",
+    "react-dom": "^16.3",
     "styled-components": "^4.x"
   },
   "nyc": {
@@ -212,8 +213,9 @@
     "instrument": false,
     "exclude": [
       "test/**/*.js",
-      "src/tasks/*.js",
-      "src/templates/*.js"
+      "src/tasks",
+      "src/templates",
+      "src/styles"
     ]
   },
   "maintainers": [
