@@ -112,6 +112,12 @@ function arrayify(thing) {
   return Array.isArray(thing) ? thing : [thing];
 }
 
+function getSampleFromOptions(options) {
+  return options.length
+    ? options[0]
+    : options;
+}
+
 // this can be moved into a action and call kepler.gl action
 /**
  *
@@ -122,13 +128,13 @@ function arrayify(thing) {
 export const loadRemoteResourceSuccess = (state, action) => {
   const responses = arrayify(action.response);
   const options = arrayify(action.options);
+  const currentSample = getSampleFromOptions(action.options);
 
   const datasets = options.map((opt, i) => {
     const datasetId = opt.id || generateHashId(6);
     const { dataUrl } = opt;
-  
+
     let processorMethod = processCsvData;
-    // TODO: create helper to determine file ext eligibility
     if (dataUrl.includes('.json') || dataUrl.includes('.geojson')) {
       processorMethod = processGeojson;
     }
@@ -159,8 +165,8 @@ export const loadRemoteResourceSuccess = (state, action) => {
     ...state,
     app: {
       ...state.app,
-      currentSample: action.options,
-      isMapLoading: false // we turn of the spinner
+      currentSample,
+      isMapLoading: false // we turn off the spinner
     },
     keplerGl: {
       ...state.keplerGl, // in case you keep multiple instances
