@@ -18,6 +18,8 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
+import {MAP_URI} from '../constants/default-settings';
+
 export function parseQueryString(query) {
   const searchParams = new URLSearchParams(query);
   const params = {};
@@ -34,8 +36,10 @@ export function parseQueryString(query) {
  * @param mapLink the cloud-providers url used to store the map
  * @returns {string}
  */
-export function getMapPermalink(mapLink) {
-  return `${window.location.protocol}//${window.location.host}/${mapLink}`
+export function getMapPermalink(mapLink, fullURL = true) {
+  return (fullURL)
+    ? `${window.location.protocol}//${window.location.host}/${MAP_URI}${mapLink}`
+    : `/${MAP_URI}${mapLink}`;
 }
 
 // from http://blog.stevenlevithan.com/archives/parseuri
