@@ -19,6 +19,7 @@
 // THE SOFTWARE.
 
 import React, {Component, createRef} from 'react';
+import {polyfill} from 'react-lifecycles-compat';
 import PropTypes from 'prop-types';
 import fuzzy from 'fuzzy';
 import classNames from 'classnames';
@@ -68,7 +69,56 @@ const InputIcon = styled.div`
   color: ${props => props.theme.inputPlaceholderColor};
 `;
 
-export default class Typeahead extends Component {
+function generateSearchFunction(props) {
+  const {searchOptions, filterOption} = props;
+  if (typeof searchOptions === 'function') {
+    if (filterOption !== null) {
+      Console.warn(
+        'searchOptions prop is being used, filterOption prop will be ignored'
+      );
+    }
+    return searchOptions;
+  } else if (typeof filterOption === 'function') {
+    // use custom filter option
+    return (value, options) => options.filter(o => filterOption(value, o));
+  }
+
+  const mapper =
+    typeof filterOption === 'string'
+      ? Accessor.generateAccessor(filterOption)
+      : Accessor.IDENTITY_FN;
+
+  return (value, options) =>
+    fuzzy
+      .filter(value, options, {extract: mapper})
+      .map(res => options[res.index]);
+}
+
+function getOptionsForValue(value, props, state) {
+  const {options, showOptionsWhenEmpty} = props;
+
+  if (!props.searchable) {
+    // directly pass through options if can not be searched
+    return options;
+  }
+  if (shouldSkipSearch(value, state, showOptionsWhenEmpty)) {
+    return options;
+  }
+
+  const searchOptions = generateSearchFunction(props);
+  return searchOptions(value, options);
+}
+
+function shouldSkipSearch(input, state, showOptionsWhenEmpty) {
+  const emptyValue = !input || input.trim().length === 0;
+
+  // this.state must be checked because it may not be defined yet if this function
+  // is called from within getInitialState
+  const isFocused = state && state.isFocused;
+  return !(showOptionsWhenEmpty && isFocused) && emptyValue;
+}
+
+class Typeahead extends Component {
   static propTypes = {
     name: PropTypes.string,
     customClasses: PropTypes.object,
@@ -96,7 +146,10 @@ export default class Typeahead extends Component {
     inputDisplayOption: PropTypes.oneOfType([PropTypes.string, PropTypes.func]),
     formInputOption: PropTypes.oneOfType([PropTypes.string, PropTypes.func]),
     defaultClassNames: PropTypes.bool,
-    customListComponent: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
+    customListComponent: PropTypes.oneOfType([
+      PropTypes.element,
+      PropTypes.func
+    ]),
     customListItemComponent: PropTypes.oneOfType([
       PropTypes.element,
       PropTypes.func
@@ -138,14 +191,18 @@ export default class Typeahead extends Component {
     resultsTruncatedMessage: null
   };
 
+  static getDerivedStateFromProps(props, state) {
+    //  invoked after a component is instantiated as well as before it is re-rendered
+    const searchResults = getOptionsForValue(state.entryValue, props, state);
+
+    return {searchResults};
+  }
+
   constructor(props) {
     super(props);
 
     this.state = {
-      searchResults: this.getOptionsForValue(
-        this.props.initialValue,
-        this.props.options
-      ),
+      searchResults: [],
 
       // This should be called something else, 'entryValue'
       entryValue: this.props.value || this.props.initialValue,
@@ -163,10 +220,6 @@ export default class Typeahead extends Component {
   }
 
   componentDidMount() {
-    this.setState({
-      searchResults: this.getOptionsForValue('', this.props.options)
-    });
-
     // call focus on entry or div to trigger key events listener
     if (this.entry.current) {
       this.entry.current.focus();
@@ -175,57 +228,26 @@ export default class Typeahead extends Component {
     }
   }
 
-  componentWillReceiveProps(nextProps) {
-    const searchResults = this.getOptionsForValue(
-      this.state.entryValue,
-      nextProps.options
-    );
-
-    this.setState({searchResults});
-  }
-
   root = createRef();
   entry = createRef();
 
-  _shouldSkipSearch(input) {
-    const emptyValue = !input || input.trim().length === 0;
-
-    // this.state must be checked because it may not be defined yet if this function
-    // is called from within getInitialState
-    const isFocused = this.state && this.state.isFocused;
-    return !(this.props.showOptionsWhenEmpty && isFocused) && emptyValue;
-  }
-
-  getOptionsForValue(value, options) {
-    if (!this.props.searchable) {
-      // directly pass through options if can not be searched
-      return options;
-    }
-    if (this._shouldSkipSearch(value)) {
-      return options;
-    }
-
-    const searchOptions = this._generateSearchFunction();
-    return searchOptions(value, options);
-  }
-
-  focus() {
+  focus = () => {
     if (this.entry.current) {
       this.entry.current.focus();
     }
-  }
+  };
 
-  _hasCustomValue() {
+  _hasCustomValue = () => {
     return (
       this.props.allowCustomValues > 0 &&
       this.state.entryValue.length >= this.props.allowCustomValues &&
       this.state.searchResults.indexOf(this.state.entryValue) < 0
     );
-  }
+  };
 
-  _getCustomValue() {
+  _getCustomValue = () => {
     return this._hasCustomValue() ? this.state.entryValue : null;
-  }
+  };
 
   _renderIncrementalSearchResults() {
     return (
@@ -276,7 +298,7 @@ export default class Typeahead extends Component {
     if (this.props.searchable) {
       // reset entry input
       this.setState({
-        searchResults: this.getOptionsForValue('', this.props.options),
+        searchResults: getOptionsForValue('', this.props, this.state),
         selection: '',
         entryValue: ''
       });
@@ -291,7 +313,7 @@ export default class Typeahead extends Component {
       const value = this.entry.current.value;
 
       this.setState({
-        searchResults: this.getOptionsForValue(value, this.props.options),
+        searchResults: getOptionsForValue(value, this.props, this.state),
         selection: '',
         entryValue: value
       });
@@ -306,19 +328,19 @@ export default class Typeahead extends Component {
     return this._onOptionSelected(selection, event);
   };
 
-  _onEscape() {
+  _onEscape = () => {
     this.setState({
       selectionIndex: null
     });
-  }
+  };
 
-  _onTab(event) {
+  _onTab = event => {
     const selection = this.getSelection();
     let option = selection
       ? selection
       : this.state.searchResults.length > 0
-        ? this.state.searchResults[0]
-        : null;
+      ? this.state.searchResults[0]
+      : null;
 
     if (option === null && this._hasCustomValue()) {
       option = this._getCustomValue();
@@ -327,9 +349,9 @@ export default class Typeahead extends Component {
     if (option !== null) {
       return this._onOptionSelected(option, event);
     }
-  }
+  };
 
-  eventMap(event) {
+  eventMap = event => {
     const events = {};
 
     events[KeyEvent.DOM_VK_UP] = this.navUp;
@@ -341,15 +363,17 @@ export default class Typeahead extends Component {
     events[KeyEvent.DOM_VK_TAB] = this._onTab;
 
     return events;
-  }
+  };
 
-  _nav(delta) {
+  _nav = delta => {
     if (!this._hasHint()) {
       return;
     }
     let newIndex =
       this.state.selectionIndex === null
-        ? delta === 1 ? 0 : delta
+        ? delta === 1
+          ? 0
+          : delta
         : this.state.selectionIndex + delta;
     let length = this.props.maxVisible
       ? this.state.searchResults.slice(0, this.props.maxVisible).length
@@ -365,7 +389,7 @@ export default class Typeahead extends Component {
     }
 
     this.setState({selectionIndex: newIndex});
-  }
+  };
 
   navDown = () => {
     this._nav(1);
@@ -430,33 +454,6 @@ export default class Typeahead extends Component {
     );
   }
 
-  _generateSearchFunction() {
-    const searchOptionsProp = this.props.searchOptions;
-    const filterOptionProp = this.props.filterOption;
-    if (typeof searchOptionsProp === 'function') {
-      if (filterOptionProp !== null) {
-        Console.warn(
-          'searchOptions prop is being used, filterOption prop will be ignored'
-        );
-      }
-      return searchOptionsProp;
-    } else if (typeof filterOptionProp === 'function') {
-      // use custom filter option
-      return (value, options) =>
-        options.filter(o => filterOptionProp(value, o));
-    }
-
-    const mapper =
-      typeof filterOptionProp === 'string'
-        ? Accessor.generateAccessor(filterOptionProp)
-        : Accessor.IDENTITY_FN;
-
-    return (value, options) =>
-      fuzzy
-        .filter(value, options, {extract: mapper})
-        .map(res => options[res.index]);
-  }
-
   _hasHint() {
     return this.state.searchResults.length > 0 || this._hasCustomValue();
   }
@@ -492,25 +489,29 @@ export default class Typeahead extends Component {
       >
         {this._renderHiddenInput()}
         {this.props.searchable ? (
-        <InputBox>
-          <TypeaheadInput
-            ref={this.entry}
-            type="text"
-            disabled={this.props.disabled}
-            {...this.props.inputProps}
-            placeholder={this.props.placeholder}
-            className={inputClassList}
-            value={this.state.entryValue}
-            onChange={this._onChange}
-            onBlur={this._onBlur}
-          />
-          <InputIcon>
-            <Search height="18px"/>
-          </InputIcon>
-        </InputBox>
+          <InputBox>
+            <TypeaheadInput
+              ref={this.entry}
+              type="text"
+              disabled={this.props.disabled}
+              {...this.props.inputProps}
+              placeholder={this.props.placeholder}
+              className={inputClassList}
+              value={this.state.entryValue}
+              onChange={this._onChange}
+              onBlur={this._onBlur}
+            />
+            <InputIcon>
+              <Search height="18px" />
+            </InputIcon>
+          </InputBox>
         ) : null}
         {this._renderIncrementalSearchResults()}
       </TypeaheadWrapper>
     );
   }
-};
+}
+
+polyfill(Typeahead);
+
+export default Typeahead;
