@@ -32,12 +32,14 @@ import {addDataToMap} from 'actions';
 // Utils
 import {getDefaultInteraction, findFieldsToShow} from 'utils/interaction-utils';
 import {
+  applyFilterFieldName,
+  applyFiltersToDatasets,
   getDefaultFilter,
-  getFilterProps,
   getFilterPlot,
   getDefaultFilterPlotType,
-  filterData,
-  isInRange
+  isInRange,
+  FILTER_UPDATER_PROPS,
+  LIMITED_FILTER_EFFECT_PROPS
 } from 'utils/filter-utils';
 import {createNewDataEntry} from 'utils/dataset-utils';
 
@@ -432,7 +434,6 @@ export function layerVisConfigChangeUpdater(state, action) {
   const {oldLayer} = action;
   const idx = state.layers.findIndex(l => l.id === oldLayer.id);
   const props = Object.keys(action.newVisConfig);
-
   const newVisConfig = {
     ...oldLayer.config.visConfig,
     ...action.newVisConfig
@@ -506,11 +507,17 @@ export function interactionConfigChangeUpdater(state, action) {
  * @param {Number} action.idx `idx` of filter to be updated
  * @param {string} action.prop `prop` of filter, e,g, `dataId`, `name`, `value`
  * @param {*} action.value new value
+ * @param {string} datasetId used when updating a prop (dataId, name) that can be linked to multiple datasets
  * @returns {Object} nextState
  * @public
  */
 export function setFilterUpdater(state, action) {
-  const {idx, prop, value} = action;
+  const {
+    idx,
+    prop,
+    value,
+    valueIndex = 0
+  } = action;
   let newState = state;
   let newFilter = {
     ...state.filters[idx],
@@ -518,80 +525,88 @@ export function setFilterUpdater(state, action) {
   };
 
   const {dataId} = newFilter;
-  if (!dataId) {
+  if (!dataId || !dataId.length) {
     return state;
   }
-  const {fields, allData} = state.datasets[dataId];
+
+  // ENsuring backward compatibility
+  const datasetIds = Array.isArray(dataId) ? dataId : [dataId];
 
   switch (prop) {
-    case 'dataId':
+    // TODO: Next PR for UI if we update dataId, we need to consider two cases:
+    // 1. dataId is empty: create a default filter
+    // 2. Add a new dataset id
+    case FILTER_UPDATER_PROPS.dataId:
       // if trying to update filter dataId. create an empty new filter
       newFilter = getDefaultFilter(dataId);
       break;
 
-    case 'name':
-      // find the field
-      const fieldIdx = fields.findIndex(f => f.name === value);
-      let field = fields[fieldIdx];
-
-      if (!field.filterProp) {
-        // get filter domain from field
-        // save filterProps: {domain, steps, value} to field, avoid recalculate
-        field = {
-          ...field,
-          filterProp: getFilterProps(allData, field)
-        };
+    case FILTER_UPDATER_PROPS.name:
+      // we are supporting the current functionality
+      // TODO: Next PR for UI filter name will only update filter name but it won't have side effects
+      // we are gonna use pair of datasets and fieldIdx to update the filter
+      const datasetId = newFilter.dataId[valueIndex];
+      const {filter: updatedFilter, dataset: newDataset} = applyFilterFieldName(
+        newFilter,
+        state.datasets[datasetId],
+        value,
+        valueIndex,
+        {mergeDomain: false}
+      );
+      if (!updatedFilter) {
+        return state;
       }
 
-      newFilter = {
-        ...newFilter,
-        ...field.filterProp,
-        name: field.name,
-        // can't edit dataId once name is selected
-        freeze: true,
-        fieldIdx
-      };
-      const enlargedFilterIdx = state.filters.findIndex(f => f.enlarged);
-      if (enlargedFilterIdx > -1 && enlargedFilterIdx !== idx) {
-        // there should be only one enlarged filter
-        newFilter.enlarged = false;
-      }
+      newFilter = updatedFilter;
 
       newState = {
         ...state,
         datasets: {
           ...state.datasets,
-          [dataId]: {
-            ...state.datasets[dataId],
-            fields: fields.map((d, i) => (i === fieldIdx ? field : d))
-          }
+          [datasetId]: newDataset
         }
       };
+
+      // only filter the current dataset
       break;
-    case 'value':
     default:
+
       break;
+  }
+
+  const enlargedFilter = state.filters.find(f => f.enlarged);
+
+  if (enlargedFilter && enlargedFilter.id !== newFilter.id) {
+    // there should be only one enlarged filter
+    newFilter.enlarged = false;
   }
 
   // save new filters to newState
   newState = {
     ...newState,
-    filters: state.filters.map((f, i) => (i === idx ? newFilter : f))
+    filters: Object.assign([...state.filters], {[idx]: newFilter})
   };
+
+  // if we are currently setting a prop that only requires to filter the current
+  // dataset we will pass only the current dataset to applyFiltersToDatasets and
+  // updateAllLayerDomainData otherwise we pass the all list of datasets as defined in dataId
+  const datasetIdsToFilter = LIMITED_FILTER_EFFECT_PROPS[prop]
+    ? [datasetIds[valueIndex]]
+    : datasetIds;
 
   // filter data
   newState = {
     ...newState,
-    datasets: {
-      ...newState.datasets,
-      [dataId]: {
-        ...newState.datasets[dataId],
-        ...filterData(allData, dataId, newState.filters)
-      }
-    }
+    datasets: applyFiltersToDatasets(
+      datasetIdsToFilter,
+      newState.datasets,
+      newState.filters
+    )
   };
 
-  newState = updateAllLayerDomainData(newState, dataId, newFilter);
+  // dataId is an array
+  // pass only the dataset we need to update
+  newState = updateAllLayerDomainData(newState, datasetIdsToFilter, newFilter);
 
   return newState;
 }
@@ -655,11 +670,14 @@ export const addFilterUpdater = (state, action) =>
  * @param {Object} action.prop
  * @param {Object} action.newConfig
  */
-export const layerColorUIChangeUpdater = (state, {oldLayer, prop, newConfig}) => {
+export const layerColorUIChangeUpdater = (
+  state,
+  {oldLayer, prop, newConfig}
+) => {
   const newLayer = oldLayer.updateLayerColorUI(prop, newConfig);
   return {
     ...state,
-    layers: state.layers.map(l => l.id === oldLayer.id ? newLayer : l)
+    layers: state.layers.map(l => (l.id === oldLayer.id ? newLayer : l))
   };
 };
 
@@ -777,13 +795,7 @@ export const removeFilterUpdater = (state, action) => {
 
   const newState = {
     ...state,
-    datasets: {
-      ...state.datasets,
-      [dataId]: {
-        ...state.datasets[dataId],
-        ...filterData(state.datasets[dataId].allData, dataId, newFilters)
-      }
-    },
+    datasets: applyFiltersToDatasets(dataId, state.datasets, newFilters),
     filters: newFilters
   };
 
@@ -908,7 +920,9 @@ export const removeDatasetUpdater = (state, action) => {
   );
 
   // remove filters
-  const filters = state.filters.filter(filter => filter.dataId !== datasetKey);
+  const filters = state.filters.filter(
+    filter => !filter.dataId.includes(datasetKey)
+  );
 
   // update interactionConfig
   let {interactionConfig} = state;
@@ -1158,6 +1172,7 @@ export const updateVisDataUpdater = (state, action) => {
     }),
     {}
   );
+
   if (!Object.keys(newDateEntries).length) {
     return state;
   }
