@@ -21,9 +21,8 @@
 import moment from 'moment';
 import {ascending, extent, histogram as d3Histogram, ticks} from 'd3-array';
 import keyMirror from 'keymirror';
-
 import {ALL_FIELD_TYPES} from 'constants/default-settings';
-import {maybeToDate, notNullorUndefined} from './data-utils';
+import {maybeToDate, notNullorUndefined, unique} from './data-utils';
 import * as ScaleUtils from './data-scale-utils';
 import {generateHashId} from './utils';
 
@@ -59,6 +58,15 @@ export const PLOT_TYPES = keyMirror({
   lineChart: null
 });
 
+export const FILTER_UPDATER_PROPS = keyMirror({
+  dataId: null,
+  name: null
+});
+
+export const LIMITED_FILTER_EFFECT_PROPS = keyMirror({
+  [FILTER_UPDATER_PROPS.name]: null
+});
+
 const SupportedPlotType = {
   [FILTER_TYPES.timeRange]: {
     default: 'histogram',
@@ -79,34 +87,136 @@ export const FILTER_COMPONENTS = {
   [FILTER_TYPES.range]: 'RangeFilter'
 };
 
+export const DEFAULT_FILTER_STRUCTURE = {
+  dataId: [], // [string]
+  freeze: false,
+  id: null,
+
+  // time range filter specific
+  fixedDomain: false,
+  enlarged: false,
+  isAnimating: false,
+  speed: 1,
+
+  // field specific
+  name: [], // string
+  type: null,
+  fieldIdx: [], // [integer]
+  domain: null,
+  value: null,
+
+  // plot
+  plotType: PLOT_TYPES.histogram,
+  yAxis: null,
+  interval: null
+};
+
+export const FILTER_ID_LENGTH = 4;
+
+/**
+ * Generates a filter with a dataset id as dataId
+ * @param {[string]} dataId
+ * @return {object} filter
+ */
 export function getDefaultFilter(dataId) {
   return {
-    // link to dataset Id
-    dataId,
-    // should allow to edit dataId
-    freeze: false,
-    id: generateHashId(4),
-
-    // time range filter specific
-    fixedDomain: false,
-    enlarged: false,
-    isAnimating: false,
-    speed: 1,
-
-    // field specific
-    name: null,
-    type: null,
-    fieldIdx: null,
-    domain: null,
-    value: null,
-
-    // plot
-    plotType: PLOT_TYPES.histogram,
-    yAxis: null,
-    interval: null
+    ...DEFAULT_FILTER_STRUCTURE,
+    // store it as dataId and it could be one or many
+    dataId: Array.isArray(dataId) ? dataId : [dataId],
+    id: generateHashId(FILTER_ID_LENGTH)
   };
 }
 
+/**
+ * Check if a filter is valid based on the given dataId
+ * @param {object} filter to validate
+ * @param {string} dataset id to validate filter against
+ * @return {boolean} true if a filter is valid, false otherwise
+ */
+export function shouldApplyFilter(filter, datasetId) {
+  const valid = Boolean(filter.dataId)
+    // datasetId is contained in dataId
+    && (
+      // filter.dataId is an array or or matches exactly the datasetId
+      (Array.isArray(filter.dataId) && filter.dataId.includes(datasetId))
+      || filter.dataId === datasetId
+    )
+    // value must be defined
+    && filter.value !== null;
+
+  return valid;
+}
+
+/**
+ * Validate saved filter config with new data,
+ * calculate domain and fieldIdx based new fields and data
+ *
+ * @param {Object} dataset
+ * @param {Object} filter - filter to be validate
+ * @return {Object | null} - validated filter
+ */
+export function validateFilterWithData(dataset, filter) {
+
+  // match filter.dataId
+  const failed = {dataset, filter: null};
+
+  const filterDataId = Array.isArray(filter.dataId) ? filter.dataId : [filter.dataId];
+
+  const filterDatasetIndex = filterDataId.indexOf(dataset.id);
+  if (filterDatasetIndex < 0) {
+    // the current filter is not mapped against the current dataset
+    return failed;
+  }
+
+  const initializeFilter = {
+    ...getDefaultFilter(filter.dataId),
+    ...filter,
+    dataId: filterDataId,
+    name: Array.isArray(filter.name) ? filter.name : [filter.name]
+  };
+
+  const fieldName = initializeFilter.name[filterDatasetIndex];
+  const {filter: updatedFilter, dataset: updatedDataset} =
+    applyFilterFieldName(initializeFilter, dataset, fieldName, filterDatasetIndex, {mergeDomain: true});
+
+  if (!updatedFilter) {
+    return failed;
+  }
+
+  updatedFilter.value = adjustValueToFilterDomain(filter.value, updatedFilter);
+  if (updatedFilter.value === null) {
+    // cannot adjust saved value to filter
+    return failed;
+  }
+
+  return {
+    filter: validateFilterYAxis(updatedFilter, updatedDataset),
+    dataset: updatedDataset
+  }
+}
+
+function validateFilterYAxis(filter, dataset) {
+  // TODO: validate yAxis against other datasets
+
+  const {fields, allData} = dataset;
+  const {yAxis} = filter;
+  // TODO: validate yAxis against other datasets
+  if (yAxis) {
+    const matchedAxis = fields.find(
+      ({name, type}) => name === yAxis.name && type === yAxis.type
+    );
+
+    filter = matchedAxis
+      ? {
+        ...filter,
+        yAxis: matchedAxis,
+        ...getFilterPlot({...filter, yAxis: matchedAxis}, allData)
+      }
+      : filter;
+  }
+
+  return filter;
+}
 /**
  * Get default filter prop based on field type
  *
@@ -115,7 +225,7 @@ export function getDefaultFilter(dataId) {
  * @returns {object} default filter
  */
 export function getFilterProps(data, field) {
-  const filterProp = {
+  const filterProps = {
     ...getFieldDomain(data, field),
     fieldType: field.type
   };
@@ -124,15 +234,15 @@ export function getFilterProps(data, field) {
     case ALL_FIELD_TYPES.real:
     case ALL_FIELD_TYPES.integer:
       return {
-        ...filterProp,
-        value: filterProp.domain,
+        ...filterProps,
+        value: filterProps.domain,
         type: FILTER_TYPES.range,
         typeOptions: [FILTER_TYPES.range]
       };
 
     case ALL_FIELD_TYPES.boolean:
       return {
-        ...filterProp,
+        ...filterProps,
         type: FILTER_TYPES.select,
         value: true
       };
@@ -140,18 +250,18 @@ export function getFilterProps(data, field) {
     case ALL_FIELD_TYPES.string:
     case ALL_FIELD_TYPES.date:
       return {
-        ...filterProp,
+        ...filterProps,
         type: FILTER_TYPES.multiSelect,
         value: []
       };
 
     case ALL_FIELD_TYPES.timestamp:
       return {
-        ...filterProp,
+        ...filterProps,
         type: FILTER_TYPES.timeRange,
         enlarged: true,
         fixedDomain: true,
-        value: filterProp.domain
+        value: filterProps.domain
       };
 
     default:
@@ -196,56 +306,53 @@ export function getFieldDomain(data, field) {
 
 /**
  * Filter data based on an array of filters
- *
- * @param {Object[]} data
- * @param {string} dataId
- * @param {Object[]} filters
- * @returns {Object[]} data
- * @returns {Number[]} filteredIndex
+ * @param {Object} dataset to perform the filter on
+ * @param {Object[]} filters list of filters to use against dataset
  */
-export function filterData(data, dataId, filters) {
-  if (!data || !dataId) {
-    // why would there not be any data? are we over doing this?
-    return {data: [], filteredIndex: []};
-  }
+export function filterData(dataset, filters) {
+  const {allData: data, fields} = dataset;
 
   if (!filters.length) {
-    return {data, filteredIndex: data.map((d, i) => i)};
+    const defaultValues = data.map((d, i) => i);
+    return {data, filteredIndex: defaultValues, filteredIndexForDomain: defaultValues};
   }
 
-  const appliedFilters = filters.filter(
-    d => d.dataId === dataId && d.fieldIdx > -1 && d.value !== null
-  );
+  const appliedFilters = filters.filter(d => shouldApplyFilter(d, dataset.id));
+
+  // Map filter against current dataset field
+  const filtersToFields = filters.reduce((acc, filter) => {
+    const fieldIndex = getDatasetFieldIndexForFilter(dataset, filter);
+
+    return {
+      ...acc,
+      ...(fieldIndex !== -1 ? {[filter.id]: fields[fieldIndex]} : {})
+    }
+  }, {
+    // [filterId]: field
+  });
 
   const [dynamicDomainFilters, fixedDomainFilters] = appliedFilters.reduce(
     (accu, f) => {
-      if (f.dataId === dataId && f.fieldIdx > -1 && f.value !== null) {
-        (f.fixedDomain ? accu[1] : accu[0]).push(f);
-      }
+      (f.fixedDomain ? accu[1] : accu[0]).push(f);
       return accu;
     },
     [[], []]
   );
-  // console.log(dynamicDomainFilters)
-  // console.log(fixedDomainFilters)
-  // we save a reference of allData index here to access dataToFeature
-  // in geojson and hexgonId layer
-  // console.time('filterData');
 
   const {filtered, filteredIndex, filteredIndexForDomain} = data.reduce(
     (accu, d, i) => {
       // generate 2 sets of
       // filter data used to calculate layer Domain
-      const matchForDomain = dynamicDomainFilters.every(filter =>
-        isDataMatchFilter(d, filter, i)
-      );
+      const matchForDomain = dynamicDomainFilters.every(filter => {
+        return isDataMatchFilter(d, filter, i, filtersToFields[filter.id]);
+      });
 
       if (matchForDomain) {
         accu.filteredIndexForDomain.push(i);
 
         // filter data for render
         const matchForRender = fixedDomainFilters.every(filter =>
-          isDataMatchFilter(d, filter, i)
+          isDataMatchFilter(d, filter, i, filtersToFields[filter.id])
         );
 
         if (matchForRender) {
@@ -259,13 +366,6 @@ export function filterData(data, dataId, filters) {
     {filtered: [], filteredIndex: [], filteredIndexForDomain: []}
   );
 
-  // console.log('data==', data.length)
-  // console.log('filtered==', filtered.length)
-  // console.log('filteredIndex==', filteredIndex.length)
-  // console.log('filteredIndexForDomain==', filteredIndexForDomain.length)
-  //
-  // console.timeEnd('filterData');
-
   return {data: filtered, filteredIndex, filteredIndexForDomain};
 }
 
@@ -275,10 +375,12 @@ export function filterData(data, dataId, filters) {
  * @param {Object[]} data
  * @param {Object} filter
  * @param {number} i
+ * @param {field} field containing values to test data against. This is used only when
+ * testing timestamp filters
  * @returns {Boolean} - whether value falls in the range of the filter
  */
-export function isDataMatchFilter(data, filter, i) {
-  const val = data[filter.fieldIdx];
+export function isDataMatchFilter(data, filter, i, field) {
+  const val = data[field.tableFieldIndex - 1];
   if (!filter.type) {
     return true;
   }
@@ -288,9 +390,10 @@ export function isDataMatchFilter(data, filter, i) {
       return isInRange(val, filter.value);
 
     case FILTER_TYPES.timeRange:
-      const timeVal = filter.mappedValue
-        ? filter.mappedValue[i]
+      const timeVal = field && field.filterProps && Array.isArray(field.filterProps.mappedValue)
+        ? field.filterProps.mappedValue[i]
         : moment.utc(val).valueOf();
+
       return isInRange(timeVal, filter.value);
 
     case FILTER_TYPES.multiSelect:
@@ -314,7 +417,6 @@ export function isDataMatchFilter(data, filter, i) {
  * @param {String} filter.type
  * @returns {*} - adjusted value to match filter or null to remove filter
  */
-
 /* eslint-disable complexity */
 export function adjustValueToFilterDomain(value, {domain, type}) {
   if (!domain || !type) {
@@ -563,3 +665,178 @@ export function getDefaultFilterPlotType(filter) {
 
   return filterPlotTypes[filter.yAxis.type] || null;
 }
+
+/**
+ * Apply a list of filters to a given dataset
+ * @param dataset
+ * @param filters
+ * @return {Object} filtered dataset
+ */
+export function applyFilterToDataset(dataset, filters) {
+  return {
+    ...dataset,
+    ...filterData(dataset, filters)
+  };
+}
+
+/**
+ *
+ * @param datasetIds list of dataset ids to be filtered
+ * @param datasets all datasets
+ * @param filters all filters to be applied to datasets
+ * @return {{[datasetId: string]: Object}} datasets - new updated datasets
+ */
+export function applyFiltersToDatasets(datasetIds, datasets, filters) {
+  const dataIds = Array.isArray(datasetIds) ? datasetIds : [datasetIds];
+  return dataIds.reduce((acc, dataIdentifier) => ({
+    ...acc,
+    [dataIdentifier]: applyFilterToDataset(datasets[dataIdentifier], filters)
+  }), datasets);
+}
+
+/**
+ * Applies a new field name value to fielter and update both filter and dataset
+ * @param filter to be applied the new field name on
+ * @param datasets
+ * @param fieldName
+ * @return {object} {filter, datasets}
+ */
+export function applyFilterFieldName(filter, dataset, fieldName, filterDatasetIndex = 0, {mergeDomain = false} = {}) {
+
+  // using filterDatasetIndex we can filter only the specified dataset
+  const {fields, allData} = dataset;
+
+  const fieldIndex = fields.findIndex(f => f.name === fieldName);
+  // if no field with same name is found, move to the next datasets
+  if (fieldIndex === -1) {
+    // throw new Error(`fieldIndex not found. Dataset must contain a property with name: ${fieldName}`);
+    return {filter: null, dataset};
+  }
+
+  const newFilter = {
+    ...filter,
+    // TODO, since we allow to add multiple fields to a filter we can no longer freeze the filter
+    freeze: true
+  };
+
+  // TODO: validate field type
+  const field = fields[fieldIndex];
+  const filterProps = field.hasOwnProperty('filterProps') ? field.filterProps : getFilterProps(allData, field);
+
+  const filterWithProps = {
+    ...(mergeDomain ? mergeFilterDomainStep(newFilter, filterProps) : {...newFilter, ...filterProps}),
+    name: Object.assign([].concat(filter.name), {[filterDatasetIndex]: field.name}),
+    fieldIdx: Object.assign([].concat(filter.fieldIdx), {[filterDatasetIndex]: field.tableFieldIndex - 1})
+  };
+
+  const fieldWithFilterProps = {
+    ...field,
+    filterProps
+  };
+
+  const newFields = fields.map((d, i) => (i === fieldIndex ? fieldWithFilterProps : d));
+
+  return {
+    filter: filterWithProps,
+    dataset: {
+      ...dataset,
+      fields: newFields
+    }
+  };
+}
+
+/**
+ * Merge one filter with other filter prop domain
+ * @param filter
+ * @param filterProps
+ * @param fieldIndex
+ * @param datasetIndex
+ * @return {*}
+ */
+/* eslint-disable complexity */
+export function mergeFilterDomainStep(filter, filterProps) {
+  if (!filter) {
+    return null;
+  }
+
+  if (!filterProps) {
+    return filter;
+  }
+
+  if (
+    (filter.fieldType && filter.fieldType !== filterProps.fieldType)
+    || !filterProps.domain
+  ) {
+    return filter;
+  }
+
+  const combinedDomain = !filter.domain ? filterProps.domain : [
+    ...(filter.domain || []),
+    ...(filterProps.domain || [])
+  ].sort((a, b) => a - b);
+
+  const newFilter = {
+    ...filter,
+    ...filterProps,
+    domain: [
+      combinedDomain[0],
+      combinedDomain[combinedDomain.length -1]
+    ]
+  };
+
+  switch (filterProps.fieldType) {
+    case ALL_FIELD_TYPES.string:
+    case ALL_FIELD_TYPES.date:
+      return {
+        ...newFilter,
+        domain: unique(combinedDomain).sort()
+      };
+
+    case ALL_FIELD_TYPES.timestamp:
+      const step = filter.step < filterProps.step ? filter.step : filterProps.step;
+
+      return {
+        ...newFilter,
+        step
+        // step: Math.min(filter.step, filterProps.step)
+      };
+    case ALL_FIELD_TYPES.real:
+    case ALL_FIELD_TYPES.integer:
+    default:
+      return newFilter;
+  }
+}
+/* eslint-enable complexity */
+
+/**
+ * Return filter dataset index from filter.dataId
+ * @param dataset
+ * @param filter
+ * @return {*}
+ */
+export function getDatasetIndexForFilter(dataset, filter) {
+  const {dataId} = filter;
+  // dataId is an array
+  return Array.isArray(dataId) ? dataId.findIndex(id => id === dataset.id)
+    // if not an array check if the current filter.dataid is equal to dataset.id
+    : dataId === dataset.id ? 0 : -1;
+}
+
+/**
+ * Return dataset field index from filter.fieldIdx
+ * The index matches the same dataset index for filter.dataId
+ * @param dataset
+ * @param filter
+ * @return {*}
+ */
+export function getDatasetFieldIndexForFilter(dataset, filter) {
+  const datasetIndex = getDatasetIndexForFilter(dataset, filter);
+  if (datasetIndex === -1) {
+    return datasetIndex;
+  }
+
+  const fieldIndex = filter.fieldIdx[datasetIndex];
+
+  return notNullorUndefined(fieldIndex) ? fieldIndex : -1;
+}
+
