@@ -57,17 +57,20 @@ test('Components -> injector -> injectComponents', t => {
   );
 
   // test if custom header is rendered
-  t.ok(wrapper.find('.my-test-header').length, 'should render custome header');
+  t.ok(wrapper.find('.my-test-header').length, 'should render custom header');
   t.end();
 });
 
 test('Components -> injector -> missing deps', t => {
   const spy = sinon.spy(Console, 'error');
 
-  const myCustomNameFactory = () => () => <div className="my-test-header-name">name</div>;
+  const myCustomNameFactory = () => () => (
+    <div className="my-test-header-name">name</div>
+  );
   const myCustomHeaderFactory = Name => () => (
     <div className="my-test-header-1">
-      <Name />smoothie
+      <Name />
+      smoothie
     </div>
   );
   myCustomHeaderFactory.deps = [myCustomNameFactory];
@@ -85,11 +88,14 @@ test('Components -> injector -> missing deps', t => {
 
   const wrapper = mount(
     <Provider store={store}>
-      <KeplerGl id="foo" mapboxApiAccessToken="smoothie_and_milkshake"/>
+      <KeplerGl id="foo" mapboxApiAccessToken="smoothie_and_milkshake" />
     </Provider>
   );
 
-  t.ok(spy.notCalled, 'Should automatically add custom deps and not call console.error');
+  t.ok(
+    spy.notCalled,
+    'Should automatically add custom deps and not call console.error'
+  );
 
   t.ok(
     wrapper.find('.my-test-header-name').length,
@@ -105,13 +111,12 @@ test('Components -> injector -> wrong factory type', t => {
   // const spy = sinon.spy(Console, 'error');
   const myCustomHeaderFactory = Name => () => (
     <div className="my-test-header-2">
-      <Name />smoothie
+      <Name />
+      smoothie
     </div>
   );
 
-  const KeplerGl = injectComponents([
-    [undefined, myCustomHeaderFactory]
-  ]);
+  const KeplerGl = injectComponents([[undefined, myCustomHeaderFactory]]);
 
   // assume instance reducer is already mounted
   const store = mockStore({
@@ -122,7 +127,7 @@ test('Components -> injector -> wrong factory type', t => {
 
   const wrapper = mount(
     <Provider store={store}>
-      <KeplerGl id="foo" mapboxApiAccessToken="smoothie_and_milkshake"/>
+      <KeplerGl id="foo" mapboxApiAccessToken="smoothie_and_milkshake" />
     </Provider>
   );
 
@@ -130,24 +135,59 @@ test('Components -> injector -> wrong factory type', t => {
   t.equal(
     spy.getCall(0).args[0],
     'Error injecting factory: ',
-    'should warn when cannot find kepler.gl state'
+    'should warn when default factory is not provided'
   );
 
   // test if custom header is rendered
-  t.ok(wrapper.find('.side-panel__panel-header').length, 'should render default header');
+  t.ok(
+    wrapper.find('.side-panel__panel-header').length,
+    'should render default header'
+  );
 
   spy.restore();
   t.end();
 });
 
-test('Components -> injector -> withState', t => {
-  const CustomHeader = ({visState}) => (
-    <div className="my-test-header-3">smoothie</div>
+test('Components -> injector -> wrong replacement type', t => {
+  const spy = sinon.spy(Console, 'error');
+  // const spy = sinon.spy(Console, 'error');
+
+  const KeplerGl = injectComponents([[PanelHeaderFactory, undefined]]);
+
+  // assume instance reducer is already mounted
+  const store = mockStore({
+    keplerGl: {
+      foo: initialCoreState
+    }
+  });
+
+  const wrapper = mount(
+    <Provider store={store}>
+      <KeplerGl id="foo" mapboxApiAccessToken="smoothie_and_milkshake" />
+    </Provider>
   );
-  const myCustomHeaderFactory = () =>
-    withState(
-      [visStateLens, mapStateLens, uiStateLens, mapStyleLens]
-    )(CustomHeader);
+
+  t.ok(spy.calledTwice, 'should call console.error twice');
+  t.equal(
+    spy.getCall(0).args[0],
+    'Error injecting replacement for: ',
+    'should warn when replace factory is not provided'
+  );
+
+  // test if custom header is rendered
+  t.ok(
+    wrapper.find('.side-panel__panel-header').length,
+    'should render default header'
+  );
+
+  spy.restore();
+  t.end();
+});
+
+test('Components -> injector -> missing dep', t => {
+  const spy = sinon.spy(Console, 'error');
+
+  const myCustomHeaderFactory = Missing => () => <Missing />;
 
   const KeplerGl = injectComponents([
     [PanelHeaderFactory, myCustomHeaderFactory]
@@ -162,39 +202,144 @@ test('Components -> injector -> withState', t => {
 
   const wrapper = mount(
     <Provider store={store}>
-      <KeplerGl id="foo" mapboxApiAccessToken="smoothie_and_milkshake"/>
+      <KeplerGl id="foo" mapboxApiAccessToken="smoothie_and_milkshake" />
+    </Provider>
+  );
+
+  t.ok(spy.calledOnce, 'should call Console.error twice');
+  t.equal(
+    spy.getCall(0).args[0],
+    `${errorMsg.incompatibleDeps(myCustomHeaderFactory, 0)}`,
+    'should warn when deps of custom factory is not provided'
+  );
+
+  t.ok(
+    wrapper.find('.factory-missing-deps'),
+    'should render missing deps component'
+  );
+  t.equal(
+    wrapper.find('.factory-missing-deps').html(),
+    '<div class="factory-missing-deps">myCustomHeaderFactory is missing dependencies</div>',
+    'should render missing deps component'
+  );
+
+  spy.restore();
+  t.end();
+});
+
+test('Components -> injector -> withState.lens', t => {
+  const CustomHeader = ({visState}) => (
+    <div className="my-test-header-3">smoothie</div>
+  );
+  const myCustomHeaderFactory = () =>
+    withState([visStateLens, mapStateLens, uiStateLens, mapStyleLens])(
+      CustomHeader
+    );
+
+  const KeplerGl = injectComponents([
+    [PanelHeaderFactory, myCustomHeaderFactory]
+  ]);
+
+  // assume instance reducer is already mounted
+  const store = mockStore({
+    keplerGl: {
+      foo: initialCoreState
+    }
+  });
+
+  const wrapper = mount(
+    <Provider store={store}>
+      <KeplerGl id="foo" mapboxApiAccessToken="smoothie_and_milkshake" />
+    </Provider>
+  );
+
+  const header = wrapper.find(CustomHeader).at(0);
+  const props = header.props();
+
+  t.ok(header, 'should render CustomHeader');
+  // test if custom header is rendered
+  t.ok(props.visState, 'should add visState to props');
+  t.ok(props.mapState, 'should add mapState to props');
+  t.ok(props.mapStyle, 'should add mapStyle to props');
+  t.ok(props.uiState, 'should add uiState to props');
+
+  t.end();
+});
+
+test('Components -> injector -> withState.mapStateToProps', t => {
+  const CustomHeader = ({visState}) => (
+    <div className="my-test-header-3">smoothie</div>
+  );
+  const myCustomHeaderFactory = () =>
+    withState([], state => ({ids: Object.keys(state)}))(CustomHeader);
+
+  const KeplerGl = injectComponents([
+    [PanelHeaderFactory, myCustomHeaderFactory]
+  ]);
+
+  // assume instance reducer is already mounted
+  const store = mockStore({
+    keplerGl: {
+      foo: initialCoreState
+    }
+  });
+
+  const wrapper = mount(
+    <Provider store={store}>
+      <KeplerGl id="foo" mapboxApiAccessToken="smoothie_and_milkshake" />
     </Provider>
   );
 
   // test if custom header is rendered
-  t.ok(
-    wrapper
-      .find(CustomHeader)
-      .at(0)
-      .props().visState,
-    'should add visState to props'
+  const header = wrapper.find(CustomHeader).at(0);
+  const props = header.props();
+
+  t.ok(header, 'should render CustomHeader');
+  t.deepEqual(props.ids, ['keplerGl'], 'should run mapStateToProps');
+
+  t.end();
+});
+
+test('Components -> injector -> actions', t => {
+  const CustomHeader = ({add}) => (
+    <div className="my-test-header-3" onClick={add}>
+      smoothie
+    </div>
   );
-  t.ok(
-    wrapper
-      .find(CustomHeader)
-      .at(0)
-      .props().mapState,
-    'should add mapState to props'
+
+  const testAction = () => ({type: 'ADD'});
+
+  const myCustomHeaderFactory = () =>
+    withState([], state => state, {add: testAction})(CustomHeader);
+
+  const KeplerGl = injectComponents([
+    [PanelHeaderFactory, myCustomHeaderFactory]
+  ]);
+
+  // assume instance reducer is already mounted
+  const store = mockStore({
+    keplerGl: {
+      foo: initialCoreState
+    }
+  });
+
+  const wrapper = mount(
+    <Provider store={store}>
+      <KeplerGl id="foo" mapboxApiAccessToken="smoothie_and_milkshake" />
+    </Provider>
   );
-  t.ok(
-    wrapper
-      .find(CustomHeader)
-      .at(0)
-      .props().mapStyle,
-    'should add mapStyle to props'
-  );
-  t.ok(
-    wrapper
-      .find(CustomHeader)
-      .at(0)
-      .props().uiState,
-    'should add uiState to props'
-  );
+
+  // test if custom header is rendered
+  const header = wrapper.find(CustomHeader).at(0);
+  const props = header.props();
+
+  t.ok(header, 'should render CustomHeader');
+  t.ok(typeof props.add === 'function', 'should add actions');
+
+  wrapper.find(CustomHeader).simulate('click');
+
+  const lastAction = store.getActions().pop();
+  t.deepEqual(lastAction, {type: 'ADD'}, 'should dispatch custom actions');
 
   t.end();
 });
