@@ -889,14 +889,14 @@ export const expectedFields1 = [
 
 export const mergedFilters = [
   {
-    dataId: '9h10t7fyb',
+    dataId: ['9h10t7fyb'],
     freeze: true,
     id: 'vxzfwyg2v',
     enlarged: false,
     isAnimating: false,
-    name: 'song_name',
+    name: ['song_name'],
     type: 'multiSelect',
-    fieldIdx: 0,
+    fieldIdx: [0],
     domain: ['2.103.2', '2.107.3', '2.116.2', '2.117.1', '3.68.4' ],
     value: ['3.68.4', '2.117.1', '2.103.2', '2.116.2'],
     fieldType: 'string',
@@ -907,14 +907,14 @@ export const mergedFilters = [
     fixedDomain: false
   },
   {
-    dataId: '9h10t7fyb',
+    dataId: ['9h10t7fyb'],
     freeze: true,
     id: 'fo9tjm2unl',
     enlarged: true,
     isAnimating: false,
-    name: 'timestamp_local',
+    name: ['timestamp_local'],
     type: 'timeRange',
-    fieldIdx: 3,
+    fieldIdx: [3],
     plotType: 'histogram',
     yAxis: null,
     interval: null,
@@ -939,15 +939,15 @@ export const mergedFilters = [
     fixedDomain: true
   },
   {
-    dataId: '9h10t7fyb',
+    dataId: ['9h10t7fyb'],
     id: 'aesy96t5',
-    name: 'type_boolean',
+    name: ['type_boolean'],
     type: 'select',
     freeze: true,
     value: false,
     enlarged: false,
     isAnimating: false,
-    fieldIdx: 10,
+    fieldIdx: [10],
     domain: [true, false],
     fieldType: 'boolean',
     plotType: 'histogram',
@@ -957,9 +957,9 @@ export const mergedFilters = [
     fixedDomain: false
   },
   {
-    dataId: '9h10t7fyb',
+    dataId: ['9h10t7fyb'],
     id: 's1bhgjt1',
-    name: 'int_range',
+    name: ['int_range'],
     type: 'range',
     value: [78, 309],
     freeze: true,
@@ -970,7 +970,7 @@ export const mergedFilters = [
     histogram: ['not tested'],
     enlargedHistogram: ['not tested'],
     isAnimating: false,
-    fieldIdx: 9,
+    fieldIdx: [9],
     domain: [78, 694],
     step: 1,
     speed: 1,
@@ -979,14 +979,14 @@ export const mergedFilters = [
     fixedDomain: false
   },
   {
-    dataId: 'v79816te8',
+    dataId: ['v79816te8'],
     freeze: true,
     id: '5nfmxjjzl',
     enlarged: false,
     isAnimating: false,
-    name: 'ZIP_CODE',
+    name: ['ZIP_CODE'],
     type: 'range',
-    fieldIdx: 2,
+    fieldIdx: [2],
     plotType: 'histogram',
     yAxis: null,
     interval: null,
